# build: optimize Docker build process